+++ src/main/java/com/example/bobfriend/controller/RecruitmentController.java
@@ -1,7 +1,9 @@
 package com.example.bobfriend.controller;
 
-import com.example.bobfriend.model.dto.RecruitmentDto;
 import com.example.bobfriend.model.dto.Condition;
+import com.example.bobfriend.model.dto.Recruitment.Create;
+import com.example.bobfriend.model.dto.Recruitment.Response;
+import com.example.bobfriend.model.dto.Recruitment.ResponseCollection;
 import com.example.bobfriend.model.exception.RecruitmentIsFullException;
 import com.example.bobfriend.model.exception.RecruitmentNotActiveException;
 import com.example.bobfriend.model.exception.RecruitmentNotFoundException;
@@ -28,7 +30,7 @@ public class RecruitmentController {
             @RequestParam(name = "type", defaultValue = "all") Condition.SearchType type,
             Condition.Search searchCondition,
             @PageableDefault(sort = {"createdAt"}, direction = Sort.Direction.ASC) Pageable pageable) {
-        Page<RecruitmentDto.ResponseList> responseDtoList = null;
+        Page<ResponseCollection> responseDtoList = null;
 
         switch (type) {
             case owned: // 자기가 작성한
@@ -61,14 +63,14 @@ public class RecruitmentController {
     @GetMapping("/{recruitmentId}")
     public ResponseEntity getRecruitment(@PathVariable Long recruitmentId)
             throws RecruitmentNotFoundException {
-        RecruitmentDto.Response recruitmentResponseDto = recruitmentService.findById(recruitmentId);
+        Response recruitmentResponseDto = recruitmentService.findById(recruitmentId);
         return ResponseEntity.ok(recruitmentResponseDto);
     }
 
     @PostMapping
     public ResponseEntity createRecruitment(
-            @Valid @RequestBody RecruitmentDto.Request recruitmentRequestDto) {
-        RecruitmentDto.Response createdRecruitment = recruitmentService.createRecruitment(recruitmentRequestDto);
+            @Valid @RequestBody Create recruitmentRequestDto) {
+        Response createdRecruitment = recruitmentService.createRecruitment(recruitmentRequestDto);
         return ResponseEntity.ok(createdRecruitment);
     }
 
@@ -89,7 +91,7 @@ public class RecruitmentController {
     @PatchMapping("/{recruitmentId}")
     public ResponseEntity joinRecruitment(@PathVariable Long recruitmentId)
             throws RecruitmentIsFullException, RecruitmentNotActiveException {
-        RecruitmentDto.Response join = recruitmentService.joinOrUnjoin(recruitmentId);
+        Response join = recruitmentService.joinOrUnjoin(recruitmentId);
         return ResponseEntity.ok(join);
     }
 
@@ -106,13 +108,13 @@ public class RecruitmentController {
             Condition.Search searchCondition,
             @PageableDefault(sort = "createdAt", direction = Sort.Direction.ASC) Pageable pageable) {
 
-        Page<RecruitmentDto.Response> searchResult = getResponses(category, searchCondition, pageable);
+        Page<Response> searchResult = getResponses(category, searchCondition, pageable);
 
         return ResponseEntity.ok(searchResult);
     }
 
-    private Page<RecruitmentDto.Response> getResponses(Condition.SearchCategory category, Condition.Search searchCondition, Pageable pageable) {
-        Page<RecruitmentDto.Response> searchResult = null;
+    private Page<Response> getResponses(Condition.SearchCategory category, Condition.Search searchCondition, Pageable pageable) {
+        Page<Response> searchResult = null;
         switch (category) {
             case place:
                 searchResult = recruitmentService.searchRestaurant(searchCondition, pageable);
